Addition to doubly linked list
This example demonstrates addition of elements into a doubly linked list. The list contains data about students (name and grade).

Addition of elements is done so that the list is sorted in an ascending order of student grades.

In main function: When the addition of elements is finished, the list is printed out.