Update doubly_linked_list
For the two given doubly linked lists (sorted in an ascending order), the function "merge" merges them into the single list, also sorted in an ascending order. In function, it is assumed lists are not empty. After the execution of the funciton, head and tail of both lists must point to the newly created list.