assignment description
This example provides an iterative implementation of a depth first search (DFS) algorithm for a binary tree. Tree is implemented as a binary search tree BST (smaller or equal element to the left, larger element to the right). 
Function “addToTree“ implements addition of elements into the BST. Structure “node” defines tree node. 
Since stack data type is used in iterative DFS algorithm implementation (see function “iterativeDFS”), there are corresponding structures and functions for working with stack defined:
-	“atom” is a stack node
-	structure “Stack” defines stack data type
-	functions: “init_stack”, “pop” and “push”
There is also a “main” function that provides user with the option for addition of elements into a tree. The main function calls function for inorder printout of the tree as well as depth first search printout of the tree.